requirements.txt
django==3.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-dotenv
requests
coreapi
drf_yasg
pytest-django
pytest
gunicorn
djongo
pymongo
flake8
black